introduces listing for events
added a GET route that acceps an optional parameter start date and end date, returning events that fall between the two. I am unsure if it should return an event that goes beyond the end date limit, but for now I am assuming that events that go beyond the end date should be excluded.